requirements.txt
asgiref==3.5.2
autopep8==1.7.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
decorator==4.4.2
Django==4.1
django-filter==22.1
djangorestframework==3.13.1
Flask==2.2.2
idna==3.3
imageio==2.21.1
imageio-ffmpeg==0.4.7
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.1
moviepy==1.0.3
numpy==1.23.2
Pillow==9.2.0
proglog==0.1.10
pycodestyle==2.9.1
pytz==2022.2.1
requests==2.28.1
sqlparse==0.4.2
toml==0.10.2
tqdm==4.64.0
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.8.1
